+++ src/encoded/commands/write_submission_spreadsheets.py
@@ -124,6 +124,7 @@ GCC_SUBMISSION_ITEMS = [
     "UnalignedReads",
     "AlignedReads",
     "VariantCalls",
+    "SupplementaryFile",
     "Software"
 ]