+++ src/encoded/item_utils/medical_history.py
@@ -1,0 +1,6 @@
+from typing import Dict, Any
+
+
+def get_donor(properties: Dict[str, Any]):
+    """Get donor from properties."""
+    return properties.get("donor","")