+++ CHANGELOG.rst
@@ -7,6 +7,13 @@ smaht-portal
 Change Log
 ----------
 
+0.193.3
+=======
+`PR 495: SN bulk donor metadata script <http://github.com/smaht-dac/smaht-portal/pull/495>`_
+
+* Creates a script `commands/create_bulk_donor_manifest.py` that can take a list of Donor accession IDs or a search query on Donor and generates the Bulk Donor Manifest file containing the full protected donor metadata for the selected donors
+
+
 0.193.2
 =======
 `PR 505: Handle Aorta correctly in tissue classification table  <https://github.com/smaht-dac/smaht-portal/pull/505>`_